+++ tests/ColumnBuilderTest.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Mssql\Tests;
 
 use PHPUnit\Framework\Attributes\DataProviderExternal;
-use Yiisoft\Db\Mssql\Column\ColumnBuilder;
 use Yiisoft\Db\Mssql\Tests\Provider\ColumnBuilderProvider;
 use Yiisoft\Db\Mssql\Tests\Support\TestTrait;
 use Yiisoft\Db\Tests\AbstractColumnBuilderTest;
@@ -16,11 +15,6 @@ use Yiisoft\Db\Tests\AbstractColumnBuilderTest;
 class ColumnBuilderTest extends AbstractColumnBuilderTest
 {
     use TestTrait;
-
-    public function getColumnBuilderClass(): string
-    {
-        return ColumnBuilder::class;
-    }
 
     #[DataProviderExternal(ColumnBuilderProvider::class, 'buildingMethods')]
     public function testBuildingMethods(